+++ Backend/PriceTag/src/main/java/com/antunmod/pricetag/repo/StoreRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 import com.antunmod.pricetag.model.database.Store;
 
 @Service
-public interface StoreRepository extends JpaRepository<Store, Byte> {
+public interface StoreRepository extends JpaRepository<Store, Short> {
 
 	@Query(value = "SELECT store_ID FROM store natural join store_location "
 			+ "WHERE store_address = ?1", nativeQuery = true)
@@ -19,7 +19,7 @@ public interface StoreRepository extends JpaRepository<Store, Byte> {
 			+ "WHERE store_name = ?1", nativeQuery = true)
 	List<String> getStoreLocations(String storeName);
 	
-	Store findByStoreId(Byte storeId);
+	Store findByStoreId(Short storeId);
 	
 	Store findByStoreName(String storeName);
 }
